Fix circular import when using run context plugins (#1490)
