+++ mlflow/tracking/context/abstract_context.py
@@ -1,0 +1,24 @@
+from abc import ABCMeta, abstractmethod
+
+
+class RunContextProvider(object):
+
+    __metaclass__ = ABCMeta
+
+    @abstractmethod
+    def in_context(self):
+        """
+        Determine if MLflow is running in this context.
+
+        :return: bool indicating if in this context
+        """
+        pass
+
+    @abstractmethod
+    def tags(self):
+        """
+        Generate context-specific tags.
+
+        :return: dict of tags
+        """
+        pass
