+++ tests/tracking/context/test_registry.py
@@ -1,115 +1,12 @@
 import mock
 import pytest
-import git
 from six.moves import reload_module as reload
 
-from mlflow.entities import SourceType
-from mlflow.utils.mlflow_tags import MLFLOW_USER, MLFLOW_SOURCE_NAME, MLFLOW_SOURCE_TYPE, \
-    MLFLOW_GIT_COMMIT, MLFLOW_DATABRICKS_NOTEBOOK_ID, MLFLOW_DATABRICKS_NOTEBOOK_PATH, \
-    MLFLOW_DATABRICKS_WEBAPP_URL
-import mlflow.tracking.context
-from mlflow.tracking.context import DefaultRunContext, GitRunContext, \
-    DatabricksNotebookRunContext, RunContextProviderRegistry, resolve_tags
-
-
-MOCK_SCRIPT_NAME = "/path/to/script.py"
-MOCK_COMMIT_HASH = "commit-hash"
-
-
-@pytest.fixture
-def patch_script_name():
-    patch_sys_argv = mock.patch("sys.argv", [MOCK_SCRIPT_NAME])
-    patch_os_path_isfile = mock.patch("os.path.isfile", return_value=False)
-    with patch_sys_argv, patch_os_path_isfile:
-        yield
-
-
-@pytest.fixture
-def patch_git_repo():
-    mock_repo = mock.Mock()
-    mock_repo.head.commit.hexsha = MOCK_COMMIT_HASH
-    with mock.patch("git.Repo", return_value=mock_repo):
-        yield mock_repo
-
-
-def test_default_run_context_in_context():
-    assert DefaultRunContext().in_context() is True
-
-
-def test_default_run_context_tags(patch_script_name):
-    mock_user = mock.Mock()
-    with mock.patch("getpass.getuser", return_value=mock_user):
-        assert DefaultRunContext().tags() == {
-            MLFLOW_USER: mock_user,
-            MLFLOW_SOURCE_NAME: MOCK_SCRIPT_NAME,
-            MLFLOW_SOURCE_TYPE: SourceType.to_string(SourceType.LOCAL)
-        }
-
-
-def test_git_run_context_in_context_true(patch_script_name, patch_git_repo):
-    assert GitRunContext().in_context()
-
-
-def test_git_run_context_in_context_false(patch_script_name):
-    with mock.patch("git.Repo", side_effect=git.InvalidGitRepositoryError):
-        assert not GitRunContext().in_context()
-
-
-def test_git_run_context_tags(patch_script_name, patch_git_repo):
-    assert GitRunContext().tags() == {
-        MLFLOW_GIT_COMMIT: MOCK_COMMIT_HASH
-    }
-
-
-def test_git_run_context_caching(patch_script_name):
-    """Check that the git commit hash is only looked up once."""
-
-    mock_repo = mock.Mock()
-    mock_hexsha = mock.PropertyMock(return_value=MOCK_COMMIT_HASH)
-    type(mock_repo.head.commit).hexsha = mock_hexsha
-
-    with mock.patch("git.Repo", return_value=mock_repo):
-        context = GitRunContext()
-        context.in_context()
-        context.tags()
-
-    assert mock_hexsha.call_count == 1
-
-
-def test_databricks_notebook_run_context_in_context():
-    with mock.patch("mlflow.utils.databricks_utils.is_in_databricks_notebook") as in_notebook_mock:
-        assert DatabricksNotebookRunContext().in_context() == in_notebook_mock.return_value
-
-
-def test_databricks_notebook_run_context_tags():
-    patch_notebook_id = mock.patch("mlflow.utils.databricks_utils.get_notebook_id")
-    patch_notebook_path = mock.patch("mlflow.utils.databricks_utils.get_notebook_path")
-    patch_webapp_url = mock.patch("mlflow.utils.databricks_utils.get_webapp_url")
-
-    with patch_notebook_id as notebook_id_mock, patch_notebook_path as notebook_path_mock, \
-            patch_webapp_url as webapp_url_mock:
-        assert DatabricksNotebookRunContext().tags() == {
-            MLFLOW_SOURCE_NAME: notebook_path_mock.return_value,
-            MLFLOW_SOURCE_TYPE: SourceType.to_string(SourceType.NOTEBOOK),
-            MLFLOW_DATABRICKS_NOTEBOOK_ID: notebook_id_mock.return_value,
-            MLFLOW_DATABRICKS_NOTEBOOK_PATH: notebook_path_mock.return_value,
-            MLFLOW_DATABRICKS_WEBAPP_URL: webapp_url_mock.return_value
-        }
-
-
-def test_databricks_notebook_run_context_tags_nones():
-    patch_notebook_id = mock.patch("mlflow.utils.databricks_utils.get_notebook_id",
-                                   return_value=None)
-    patch_notebook_path = mock.patch("mlflow.utils.databricks_utils.get_notebook_path",
-                                     return_value=None)
-    patch_webapp_url = mock.patch("mlflow.utils.databricks_utils.get_webapp_url",
-                                  return_value=None)
-
-    with patch_notebook_id, patch_notebook_path, patch_webapp_url:
-        assert DatabricksNotebookRunContext().tags() == {
-            MLFLOW_SOURCE_NAME: None,
-            MLFLOW_SOURCE_TYPE: SourceType.to_string(SourceType.NOTEBOOK),
-        }
+import mlflow.tracking.context.registry
+from mlflow.tracking.context.default_context import DefaultRunContext
+from mlflow.tracking.context.git_context import GitRunContext
+from mlflow.tracking.context.databricks_notebook_context import DatabricksNotebookRunContext
+from mlflow.tracking.context.registry import RunContextProviderRegistry, resolve_tags
 
 
 def test_run_context_provider_registry_register():
@@ -158,7 +55,7 @@ def test_run_context_provider_registry_register_entrypoints_handles_exception(ex
 def _currently_registered_run_context_provider_classes():
     return {
         provider.__class__
-        for provider in mlflow.tracking.context._run_context_provider_registry
+        for provider in mlflow.tracking.context.registry._run_context_provider_registry
     }
 
 
@@ -180,7 +77,7 @@ def test_registry_instance_loads_entrypoints():
     ) as mock_get_group_all:
         # Entrypoints are registered at import time, so we need to reload the module to register th
         # entrypoint given by the mocked extrypoints.get_group_all
-        reload(mlflow.tracking.context)
+        reload(mlflow.tracking.context.registry)
 
     assert MockRunContext in _currently_registered_run_context_provider_classes()
     mock_get_group_all.assert_called_once_with("mlflow.run_context_provider")
@@ -190,7 +87,7 @@ def test_registry_instance_loads_entrypoints():
 def test_run_context_provider_registry_with_installed_plugin(tmp_wkdir):
     """This test requires the package in tests/resources/mlflow-test-plugin to be installed"""
 
-    reload(mlflow.tracking.context)
+    reload(mlflow.tracking.context.registry)
 
     from mlflow_test_plugin import PluginRunContextProvider
     assert PluginRunContextProvider in _currently_registered_run_context_provider_classes()
@@ -217,7 +114,7 @@ def mock_run_context_providers():
 
     providers = [base_provider, skipped_provider, override_provider]
 
-    with mock.patch("mlflow.tracking.context._run_context_provider_registry", providers):
+    with mock.patch("mlflow.tracking.context.registry._run_context_provider_registry", providers):
         yield
 
     skipped_provider.tags.assert_not_called()
